+++ src/hlsync/__init__.py
@@ -1,3 +1,3 @@
 """Hook Line Sync."""
 
-__version__ = "0.9.14.1"
+__version__ = "0.9.14.2"